INPAY-122 - Updated InPost Pay module to recent 2.7.0 version and fixed version provider to use composer.lock instead of composer.json file

+++ packages/magento2-module-inpost-pay/src/Observer/Customer/DeleteBasketBidingApiCookieAfterCustomerLogout.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace InPost\InPostPay\Observer\Customer;
 
+use InPost\InPostPay\Provider\Config\GeneralConfigProvider;
 use InPost\InPostPay\Service\Cart\BasketBindingApiKeyCookieService;
 use Magento\Framework\Event\Observer as EventObserver;
 use Magento\Framework\Event\ObserverInterface;
@@ -12,9 +13,11 @@ class DeleteBasketBidingApiCookieAfterCustomerLogout implements ObserverInterfac
 {
     /**
      * @param BasketBindingApiKeyCookieService $basketBindingApiKeyCookieService
+     * @param GeneralConfigProvider $generalConfigProvider
      */
     public function __construct(
-        private readonly BasketBindingApiKeyCookieService $basketBindingApiKeyCookieService
+        private readonly BasketBindingApiKeyCookieService $basketBindingApiKeyCookieService,
+        private readonly GeneralConfigProvider $generalConfigProvider
     ) {
     }
 
@@ -25,6 +28,15 @@ class DeleteBasketBidingApiCookieAfterCustomerLogout implements ObserverInterfac
      */
     public function execute(EventObserver $observer): void
     {
+        if (!$this->canSync()) {
+            return;
+        }
+
         $this->basketBindingApiKeyCookieService->deleteBasketBindingKeyCookie();
+    }
+
+    private function canSync(): bool
+    {
+        return $this->generalConfigProvider->isEnabled();
     }
 }
